# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Staff Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: department_id
          in: query
          required: false
          schema:
            type: integer
        - name: role_id
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of users with roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserWithRoles"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user/employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithRoles"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithRoles"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user/employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithRoles"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}/roles:
    get:
      operationId: getUserRoles
      tags: [users]
      summary: Get roles assigned to a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Roles list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: assignRole
      tags: [users]
      summary: Assign a role to a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleAssignment"
      responses:
        "200":
          description: Role assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithRoles"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}/roles/{roleId}:
    delete:
      operationId: removeRole
      tags: [users]
      summary: Remove a role from a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Role removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithRoles"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}/signature:
    post:
      operationId: uploadSignature
      tags: [users]
      summary: Upload a signature image for a user (PNG only, max 1 MB)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/SignatureImageInput"
      responses:
        "200":
          description: Signature uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithRoles"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Roles ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Role:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
...
    RoleAssignment:
      type: object
      required: [role_id]
      properties:
        role_id:
          type: integer
    UserWithRoles:
      type: object
      required: [id, full_name, username, email, created_at, updated_at, roles]
      properties:
        id:
          type: integer
        full_name:
          type: string
        username:
          type: string
        email:
          type: string
        department_id:
          type: ["integer", "null"]
        department_name:
          type: ["string", "null"]
        signature_image:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
    UserInput:
      type: object
      required: [full_name, username, email, password]
      properties:
        full_name:
          type: string
          minLength: 1
          maxLength: 150
        username:
          type: string
          minLength: 1
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        department_id:
          type: ["integer", "null"]
        role_ids:
          type: array
          items:
            type: integer
    UserUpdate:
      type: object
      properties:
        full_name:
          type: string
          minLength: 1
          maxLength: 150
        username:
          type: string
          minLength: 1
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        department_id:
          type: ["integer", "null"]
        role_ids:
          type: array
          items:
            type: integer
...
    SignatureImageInput:
      type: object
      required: [signature]
      properties:
        signature:
          type: string
          format: binary
...